Add support for ArrayBuffer, Blob, and TypedArray
Adds support for ArrayBuffer, Blob, and most kinds of TypedArrays
to all drivers. Relies on some hairy code (that could be further refactored)
for localStorage and WebSQL, but the point is everything checks out.

PhantomJS seems to have problems with the Blob test; while I'm hesitant to
skip tests, it DOES work fine in Safari, which is the target there, so I'm
going to look into it but commit this work for now, as it's very useful.

* Fix #28
* Fix #30
* Fix #40
* Fix #63

This partially addresses #69, as it allows for Blob storage, but doesn't
address the issue of extremely large files, which neither localStorage nor
WebSQL will support (their limits are too low).