fix for issue #365
When I am running my qunit tests using karma I am getting this error when loading localforage library:

ReferenceError: Can't find variable: require

The issue is that qunit defines "module" and "module.exports" so this check we have passes (but require is undefined)

 if (typeof module !== 'undefined' && module.exports)
I have a fix for it, changing it to:

if (typeof module !== 'undefined' && module.exports && typeof require !== 'undefined')
fixes the issue. There might be a different fix when I set up karma or qunit but I didn't find one, and it might be good to check for require before using it.