Fixed a few bugs, that manifest themselves when not enough space is available.
There is a pattern in the code that calls a function which should be
asynchronous and after that it sets up the handlers. If the function
returns before those handlers are set, the code breaks. This commit fixes
one instance in the code (namely in setItem), but are (might be) more.

There is an issue with indexedDb.open, which returns request, that has
the success and error handlers. However, this function can also fail
before the handlers are set. That is why I added check after setting
the handlers, to see, if the object isn't already in error state.