chore: Drop the package-lock
Using package-lock.json in libraries is considered a bad practice since it isn't included in the published package. As a result the code that we run on our ci might end up not being the same as the code that a fresh npm install will be using. This currently only applies for `lie`'s `immediate` dependency, but will also help us pick up the latest bugfixes from our dev dependencies and possibly get rid of the GitHub security warning for our dev dependencies.